Alpha Parity Case study project
CFTC Commitment of Traders Report (CTR) is a weekly report which breaks
down the open interest in the commodity market and shows what position
producers/money manager/swap dealer holds in a particular commodity. In
this case study, we would like to look at the commitment of traders
report for soybean and soybean oil.